Process改为由Arc管理；修复Heap大小的错误（改为以u8为单位）；修复内核上下文中sstatus.spie未清零的错误

+++ los/src/mem/allocator/heap.rs
@@ -5,7 +5,7 @@ use spin::Mutex;
 use crate::config::HEAP_SIZE;
 
 #[link_section = ".data"]
-pub(crate) static HEAP: [usize; HEAP_SIZE] = [0; HEAP_SIZE];
+pub(crate) static HEAP: [u8; HEAP_SIZE] = [0; HEAP_SIZE];
 
 pub struct BuddyAllocator {
     inner: Mutex<Heap>,
